+++ src/main.js
@@ -8,69 +8,47 @@ import Sorting from './components/sorting';
 import TripDays from './components/tripDays';
 import Day from './components/day';
 import Event from './components/event';
-import Form from './components/editEvent';
+import NoPoints from './components/noPoints';
 
 
-const points = generatePoints(10);
-const daysEvents = generateDays(points);
-const total = getTotalPrice(daysEvents);
+const points = generatePoints(20);
+
 const [menuTitle, filterTitle] = document.querySelector(`.trip-controls`).children;
 const trip = document.querySelector(`.trip-info`);
-
-render(trip, new TripRoute(daysEvents).getElement(), RenderPosition.BEFOREEND);
-render(trip, new Total(total).getElement(), RenderPosition.BEFOREEND);
+const trips = document.querySelector(`.trip-events`);
 render(menuTitle, new Menu().getElement(), RenderPosition.AFTERNODE);
 render(filterTitle, new Filters().getElement(), RenderPosition.AFTERNODE);
 
-const trips = document.querySelector(`.trip-events`);
-render(trips, new Sorting().getElement(), RenderPosition.BEFOREEND);
-render(trips, new TripDays().getElement(), RenderPosition.BEFOREEND);
+if (points.length === 0) {
 
-const days = document.querySelector(`.trip-days`);
+  render(trip, new Total(0).getElement(), RenderPosition.BEFOREEND);
+  render(trips, new NoPoints().getElement(), RenderPosition.AFTERNODE);
 
-const daysElements = [];
+} else {
 
-daysEvents.map((item) => {
-  const dayDOM = new Day(item);
-  daysElements.push(dayDOM);
-  render(days, dayDOM.getElement(), RenderPosition.BEFOREEND);
-});
+  const daysEvents = generateDays(points);
+  const total = getTotalPrice(daysEvents);
 
-const renderEvent = (event, position) => {
-  const eventCard = new Event(event);
-  const eventForm = new Form(event);
-  const showButton = eventCard.getElement().querySelector(`.event__rollup-btn`);
-  const collapseButton = eventForm.getElement().querySelector(`.event__rollup-btn`);
-  const saveButton = eventForm.getElement().querySelector(`.event__save-btn`);
-  const replaceWith = (component, newComponent) => {
-    component.getElement().replaceWith(newComponent.getElement());
-  };
+  render(trip, new TripRoute(daysEvents).getElement(), RenderPosition.BEFOREEND);
+  render(trip, new Total(total).getElement(), RenderPosition.BEFOREEND);
 
-  const onEscKeyDown = (evt) => {
-    const isEscKey = evt.key === `Escape` || evt.key === `Esc`;
-    if (isEscKey) {
-      replaceWith(eventForm, eventCard);
-      document.removeEventListener(`keydown`, onEscKeyDown);
-    }
-  };
+  render(trips, new Sorting().getElement(), RenderPosition.BEFOREEND);
+  render(trips, new TripDays().getElement(), RenderPosition.BEFOREEND);
 
-  showButton.addEventListener(`click`, () => {
-    replaceWith(eventCard, eventForm);
-    document.addEventListener(`keydown`, onEscKeyDown);
+  const days = document.querySelector(`.trip-days`);
+
+  const daysElements = [];
+
+  daysEvents.map((item) => {
+    const dayDOM = new Day(item);
+    daysElements.push(dayDOM);
+    render(days, dayDOM.getElement(), RenderPosition.BEFOREEND);
   });
 
-  collapseButton.addEventListener(`click`, () => {
-    replaceWith(eventForm, eventCard);
+  daysElements.map((element) => {
+    const dayList = element.getElement().querySelector(`.trip-events__list`);
+    element.points.map((point) => Event.renderEvent(point, dayList));
   });
+}
 
-  saveButton.addEventListener(`click`, () => {
-    replaceWith(eventForm, eventCard);
-  });
 
-  render(position, eventCard.getElement(), RenderPosition.BEFOREEND);
-};
-
-daysElements.map((element) => {
-  const dayList = element.getElement().querySelector(`.trip-events__list`);
-  element.points.map((point) => renderEvent(point, dayList));
-});
